web-animations/keyframe-effect/keyframe-effect-spacing.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>KeyframeEffect spacing attribute test</title>
<meta name="assert" content="The spacing mode to use for this animation effect">
<link rel="help" href="http://w3c.github.io/web-animations/#the-keyframeeffect-interface">
<link rel="help" href="http://w3c.github.io/web-animations/#spacing-keyframes">
<link rel="help" href="http://w3c.github.io/web-animations/#applying-spacing-to-keyframes">
<link rel="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
<body>
<div id="log"></div>
<script>
// Creates two effects that tested in the tests below. One with spacing,
// defined by the constructor. And the second with spacing set by spacing attribute
function createEffectsToTest(keyframes, spacingMode) {
    var effect1 = new KeyframeEffect(keyframes, {spacing: spacingMode});
    var effect2 = new KeyframeEffect(keyframes);
    effect2.spacing = spacingMode;
    return [effect1, effect2];
}

test(function() {
    var effect = new KeyframeEffect([]);
    assert_equals(effect.spacing, 'distribute', 'Default value of spacing should be ' +
        '\'distribute\'');
}, 'Test that KeyframeEffect constructor by default creates keyframe effect with ' +
    '\'distribute\' spacing mode');

test(function() {
    var effectsToTest = createEffectsToTest([], 'paced');

    effectsToTest.forEach(function(effect, index) {
        assert_equals(effect.spacing, 'paced', 'Effect ' + index + ': ' +
            'Spacing attribute value should be set');
    });
}, 'Test that spacing attribute returns the current spacing mode');

test(function() {
    var keyframe = {top: '10px'};
    var effectsToTest = createEffectsToTest(keyframe, 'distribute');

    effectsToTest.forEach(function(effect, index) {
        assert_equals(effect.getFrames()[0].computedOffset, 1, 'Effect ' + index + ': ' +
            'Wrong computedOffset value');
    });
}, 'Test that spacing of a single keyframe with spacing mode \'distribute\' ' +
    'produces computedOffset 1');

test(function() {
    var keyframe = {top: '10px'};
    var effectsToTest = createEffectsToTest(keyframe, 'paced');

    effectsToTest.forEach(function(effect, index) {
        assert_equals(effect.getFrames()[0].computedOffset, 1, 'Effect ' + index + ': ' +
            'Wrong computedOffset value');
    });
}, 'Test that spacing of a single keyframe with spacing mode \'paced\' ' +
    'produces computedOffset 1');

test(function() {
    var keyframe = {top: '10px'};
    var effectsToTest = createEffectsToTest(keyframe, 'paced(top)');

    effectsToTest.forEach(function(effect, index) {
        assert_equals(effect.getFrames()[0].computedOffset, 1, 'Effect ' + index + ': ' +
            'Wrong computedOffset value');
    });
}, 'Test that spacing of a single keyframe with spacing mode \'paced({ident})\' ' +
    'produces computedOffset 1');

test(function() {
    var keyframes = [{top: '10px'},
        {top: '20px'},
        {top: '50px'}];
    var effectsToTest = createEffectsToTest(keyframes, 'distribute');
    var expected = computeOffset(keyframes, 'distribute');

    effectsToTest.forEach(function(effect, index) {
        expected.forEach(function(keyframe, i) {
            testKeyframe(effect.getFrames()[i], keyframe, 'Effect ' + index + ', frame ' + i);
        });
    });
}, 'Test spacing in \'distribute\' mode a sequence of keyframes with null offset');

test(function() {
    var keyframes = [{top: '10px'},
        {top: '20px'},
        {top: '50px'}];
    var effectsToTest = createEffectsToTest(keyframes, 'paced');
    var expected = computeOffset(keyframes, 'paced');

    effectsToTest.forEach(function(effect, index) {
        expected.forEach(function(keyframe, i) {
            testKeyframe(effect.getFrames()[i], keyframe, 'Effect ' + index + ', frame ' + i);
        });
    });
}, 'Test spacing in \'paced\' mode a sequence of keyframes with null offset');

test(function() {
    var keyframes = [{top: '10px'},
        {top: '20px'},
        {top: '50px'}];
    var effectsToTest = createEffectsToTest(keyframes, 'paced(top)');
    var expected = computeOffset(keyframes, 'paced', 'top');

    effectsToTest.forEach(function(effect, index) {
        expected.forEach(function(keyframe, i) {
            testKeyframe(effect.getFrames()[i], keyframe, 'Effect ' + index + ', frame ' + i);
        });
    });
}, 'Test spacing in \'paced({ident})\' mode a sequence of keyframes with null offset');

test(function() {
    var keyframes = [{top: '10px'},
        {top: '20px'},
        {top: '30px', offset: 0.5},
        {top: '40px'},
        {top: '50px'},
        {top: '60px'},
        {top: '70px'}];
    var effectsToTest = createEffectsToTest(keyframes, 'distribute');
    var expected = computeOffset(keyframes, 'distribute', 'top');

    effectsToTest.forEach(function(effect, index) {
        expected.forEach(function(keyframe, i) {
            testKeyframe(effect.getFrames()[i], keyframe, 'Effect ' + index + ', frame ' + i);
        });
    });
}, 'Test spacing in \'distribute\' mode when there are intermediate keyframes with ' +
    'non-null offsets in the keyframes sequence');

test(function() {
    var keyframes = [{top: '10px', offset: 0.3},
        {top: '20px'},
        {top: '30px', offset: 0.5},
        {top: '40px'},
        {top: '50px'},
        {top: '60px', offset: 0.7},
        {top: '70px'},
        {top: '80px'},
        {top: '90px'}];
    var effectsToTest = createEffectsToTest(keyframes, 'distribute');
    var expected = computeOffset(keyframes, 'distribute', 'top');

    effectsToTest.forEach(function(effect, index) {
        expected.forEach(function(keyframe, i) {
            testKeyframe(effect.getFrames()[i], keyframe, 'Effect ' + index + ', frame ' + i);
        });
    });
}, 'Test spacing in \'distribute\' mode when there are intermediate keyframes with ' +
    'non-null offsets including the first one in the keyframes sequence');

test(function() {
    var keyframes = [{top: '10px'},
        {top: '20px'},
        {top: '30px', offset: 0.5},
        {top: '40px'},
        {top: '50px'},
        {top: '60px', offset: 0.7},
        {top: '70px'},
        {top: '80px'},
        {top: '90px', offset: 0.9}];
    var effectsToTest = createEffectsToTest(keyframes, 'distribute');
    var expected = computeOffset(keyframes, 'distribute', 'top');

    effectsToTest.forEach(function(effect, index) {
        expected.forEach(function(keyframe, i) {
            testKeyframe(effect.getFrames()[i], keyframe, 'Effect ' + index + ', frame ' + i);
        });
    });
}, 'Test spacing in \'distribute\' mode when there are intermediate keyframes with ' +
    'non-null offsets including the first and the last ones in the keyframes sequence');

test(function() {
    var keyframes = [{top: '10px'},
        {top: '20px'},
        {top: '30px', offset: 0.5},
        {top: '40px'},
        {top: '50px'},
        {top: '60px', offset: 0.7},
        {top: '70px'},
        {top: '80px'},
        {top: '90px', offset: 0.9}];
    var effectsToTest = createEffectsToTest(keyframes, 'distribute');
    var expected = computeOffset(keyframes, 'distribute', 'top');

    effectsToTest.forEach(function(effect, index) {
        expected.forEach(function(keyframe, i) {
            testKeyframe(effect.getFrames()[i], keyframe, 'Effect ' + index + ', frame ' + i);
        });
    });
}, 'Test spacing in \'distribute\' mode when there are intermediate keyframes with ' +
    'non-null offsets including the last one in the keyframes sequence');

test(function() {
    var keyframes = [{top: '10px'},
        {top: '20px'},
        {top: '30px', offset: 0.5},
        {top: '40px'},
        {top: '50px'},
        {top: '160px'},
        {top: '170px'}];
    var effectsToTest = createEffectsToTest(keyframes, 'paced');
    var expected = computeOffset(keyframes, 'paced', 'top');

    effectsToTest.forEach(function(effect, index) {
        expected.forEach(function(keyframe, i) {
            testKeyframe(effect.getFrames()[i], keyframe, 'Effect ' + index + ', frame ' + i);
        });
    });
}, 'Test spacing in \'paced\' mode when there are intermediate keyframes with ' +
    'non-null offsets in the keyframes sequence');

test(function() {
    var keyframes = [{top: '10px', offset: 0.3},
        {top: '20px'},
        {top: '30px', offset: 0.5},
        {top: '40px'},
        {top: '150px'},
        {top: '60px', offset: 0.7},
        {top: '70px'},
        {top: '180px'},
        {top: '190px'}];
    var effectsToTest = createEffectsToTest(keyframes, 'paced');
    var expected = computeOffset(keyframes, 'paced', 'top');

    effectsToTest.forEach(function(effect, index) {
        expected.forEach(function(keyframe, i) {
            testKeyframe(effect.getFrames()[i], keyframe, 'Effect ' + index + ', frame ' + i);
        });
    });
}, 'Test spacing in \'paced\' mode when there are intermediate keyframes with ' +
    'non-null offsets including the first one in the keyframes sequence');

test(function() {
    var keyframes = [{top: '10px'},
        {top: '20px'},
        {top: '30px', offset: 0.5},
        {top: '40px'},
        {top: '150px'},
        {top: '60px', offset: 0.7},
        {top: '70px'},
        {top: '180px'},
        {top: '190px', offset: 0.9}];
    var effectsToTest = createEffectsToTest(keyframes, 'paced');
    var expected = computeOffset(keyframes, 'paced', 'top');

    effectsToTest.forEach(function(effect, index) {
        expected.forEach(function(keyframe, i) {
            testKeyframe(effect.getFrames()[i], keyframe, 'Effect ' + index + ', frame ' + i);
        });
    });
}, 'Test spacing in \'paced\' mode when there are intermediate keyframes with ' +
    'non-null offsets including the first and the last ones in the keyframes sequence');

test(function() {
    var keyframes = [{top: '10px'},
        {top: '20px'},
        {top: '30px', offset: 0.5},
        {top: '40px'},
        {top: '150px'},
        {top: '60px', offset: 0.7},
        {top: '70px'},
        {top: '180px'},
        {top: '190px', offset: 0.9}];
    var effectsToTest = createEffectsToTest(keyframes, 'paced');
    var expected = computeOffset(keyframes, 'paced', 'top');

    effectsToTest.forEach(function(effect, index) {
        expected.forEach(function(keyframe, i) {
            testKeyframe(effect.getFrames()[i], keyframe, 'Effect ' + index + ', frame ' + i);
        });
    });
}, 'Test spacing in \'paced\' mode when there are intermediate keyframes with ' +
    'non-null offsets including the last one in the keyframes sequence');

test(function() {
    var keyframes = [{width: '20px', top: '10px'},
        {top: '20px'},
        {top: '30px', offset: 0.5},
        {top: '40px'},
        {top: '50px', width: '100px', offset: 0.7},
        {top: '160px', width: '200px'},
        {top: '170px'}];
    var effectsToTest = createEffectsToTest(keyframes, 'paced');
    var expected = computeOffset(keyframes, 'paced', 'top');

    effectsToTest.forEach(function(effect, index) {
        expected.forEach(function(keyframe, i) {
            testKeyframe(effect.getFrames()[i], keyframe, 'Effect ' + index + ', frame ' + i);
        });
    });
}, 'Test that the paced property to use is the first property specified ' +
    'in the first keyframe in the list of keyframes associated with this ' +
    'animation effect when sorting the CSS property names in ascending ' +
    'order by Unicode codepoint');

test(function() {
    var keyframes = [{width: '20px', top: '10px'},
        {top: '20px'},
        {top: '30px', offset: 0.5},
        {top: '40px'},
        {top: '50px', width: '100px', offset: 0.7},
        {top: '160px', width: '200px'},
        {top: '170px'}];
    var effectsToTest = createEffectsToTest(keyframes, 'paced(width)');
    var expected = computeOffset(keyframes, 'paced', 'width');

    effectsToTest.forEach(function(effect, index) {
        expected.forEach(function(keyframe, i) {
            testKeyframe(effect.getFrames()[i], keyframe, 'Effect ' + index + ', frame ' + i);
        });
    });
}, 'Test that spacing in paced({ident}) mode uses the paced keyframe ' +
    'spacing mode with the property name indicated by {ident} as the paced property');

test(function() {
    var keyframes = [{top: '10px', width: '20px'},
        {top: '20px'},
        {top: '30px'},
        {top: '40px'},
        {top: '50px', offset: 0.5},
        {top: '60px'},
        {top: '70px'},
        {top: '80px', width: '30px'},
        {top: '90px'},
        {top: '100px'},
        {top: '110px', width: '50px'},
        {top: '120px'},
        {top: '130px'},
        {top: '140px', width: '100px', offset: 0.7},
        {top: '150px', width: '200px'},
        {top: '160px', width: '400px'},
        {top: '170px', width: '410px'},
        {top: '180px', width: '450px', offset: 0.8},
        {top: '190px', width: '470px'},
        {top: '200px'},
        {top: '210px', width: '480px'},
        {top: '220px'}];
    var effectsToTest = createEffectsToTest(keyframes, 'paced(width)');
    var expected = computeOffset(keyframes, 'paced', 'width');

    effectsToTest.forEach(function(effect, index) {
        expected.forEach(function(keyframe, i) {
            testKeyframe(effect.getFrames()[i], keyframe, 'Effect ' + index + ', frame ' + i);
        });
    });
}, 'Test that spacing in paced({ident}) mode when there are different combinations of ' +
    'paceble and not paceble keyframes');

test(function() {
    var keyframes = [{top: '0px'}, // A
        {top: '10px'},
        {top: '20px'},
        {top: '30px', width: '30px'}, // pacedA
        {top: '40px'},
        {top: '50px'},
        {top: '60px', width: '50px'},
        {top: '70px'},
        {top: '80px'},
        {top: '90px', width: '100px'}, // pacedB
        {top: '100px'},
        {top: '110px'},
        {top: '120px'}]; //B
    var effectsToTest = createEffectsToTest(keyframes, 'paced(width)');
    var expected = computeOffset(keyframes, 'paced', 'width');

    effectsToTest.forEach(function(effect, index) {
        expected.forEach(function(keyframe, i) {
            testKeyframe(effect.getFrames()[i], keyframe, 'Effect ' + index + ', frame ' + i);
        });
    });
}, 'Test that spacing in paced mode algorithm: there are frames between A, pacedA, pacedB ' +
    'and B');

test(function() {
    var keyframes = [{top: '0px'}, // A
        {top: '10px', width: '30px'}, // pacedA
        {top: '20px'},
        {top: '30px'},
        {top: '40px', width: '50px'},
        {top: '50px'},
        {top: '60px'},
        {top: '70px', width: '100px'}, // pacedB
        {top: '80px'},
        {top: '90px'},
        {top: '100px'}]; //B
    var effectsToTest = createEffectsToTest(keyframes, 'paced(width)');
    var expected = computeOffset(keyframes, 'paced', 'width');

    effectsToTest.forEach(function(effect, index) {
        expected.forEach(function(keyframe, i) {
            testKeyframe(effect.getFrames()[i], keyframe, 'Effect ' + index + ', frame ' + i);
        });
    });
}, 'Test that spacing in paced mode algorithm: no frames between A and pacedA');

test(function() {
    var keyframes = [{top: '0px', width: '10px'}, // A, pacedA
        {top: '10px'},
        {top: '20px'},
        {top: '30px', width: '30px'},
        {top: '40px'},
        {top: '50px'},
        {top: '60px', width: '50px'},
        {top: '70px'},
        {top: '80px'},
        {top: '90px', width: '100px'}, // pacedB
        {top: '100px'},
        {top: '110px'},
        {top: '120px'}]; //B
    var effectsToTest = createEffectsToTest(keyframes, 'paced(width)');
    var expected = computeOffset(keyframes, 'paced', 'width');

    effectsToTest.forEach(function(effect, index) {
        expected.forEach(function(keyframe, i) {
            testKeyframe(effect.getFrames()[i], keyframe, 'Effect ' + index + ', frame ' + i);
        });
    });
}, 'Test that spacing in paced mode algorithm: A == pacedA');

test(function() {
    var keyframes = [{top: '0px'}, // A
        {top: '10px'},
        {top: '20px'},
        {top: '30px', width: '30px'}, // pacedA
        {top: '40px', width: '100px'}, // pacedB
        {top: '50px'},
        {top: '60px'},
        {top: '70px'}]; //B
    var effectsToTest = createEffectsToTest(keyframes, 'paced(width)');
    var expected = computeOffset(keyframes, 'paced', 'width');

    effectsToTest.forEach(function(effect, index) {
        expected.forEach(function(keyframe, i) {
            testKeyframe(effect.getFrames()[i], keyframe, 'Effect ' + index + ', frame ' + i);
        });
    });
}, 'Test that spacing in paced mode algorithm: there are no frames between pacedA and pacedB');

test(function() {
    var keyframes = [{top: '0px'}, // A
        {top: '10px'},
        {top: '20px'},
        {top: '30px', width: '100px'},
        {top: '40px'},
        {top: '50px'},
        {top: '60px'}]; //pacedA, pacedB, B
    var effectsToTest = createEffectsToTest(keyframes, 'paced(width)');
    var expected = computeOffset(keyframes, 'paced', 'width');

    effectsToTest.forEach(function(effect, index) {
        expected.forEach(function(keyframe, i) {
            testKeyframe(effect.getFrames()[i], keyframe, 'Effect ' + index + ', frame ' + i);
        });
    });
}, 'Test that spacing in paced mode algorithm: only one paceble frame ' +
    'between A and B - spacing behavior degenerates to distribute spacing');

test(function() {
    var keyframes = [{top: '0px'}, // A
        {top: '10px'},
        {top: '20px'},
        {top: '30px', width: '30px'}, // pacedA
        {top: '40px'},
        {top: '50px'},
        {top: '60px', width: '50px'},
        {top: '70px'},
        {top: '80px'},
        {top: '90px', width: '100px'}, // pacedB
        {top: '100px'}]; //B
    var effectsToTest = createEffectsToTest(keyframes, 'paced(width)');
    var expected = computeOffset(keyframes, 'paced', 'width');

    effectsToTest.forEach(function(effect, index) {
        expected.forEach(function(keyframe, i) {
            testKeyframe(effect.getFrames()[i], keyframe, 'Effect ' + index + ', frame ' + i);
        });
    });
}, 'Test that spacing in paced mode algorithm: there are no frames between pacedB and B');

test(function() {
    var keyframes = [{top: '0px'}, // A
        {top: '10px'},
        {top: '20px'},
        {top: '30px', width: '30px'}, // pacedA
        {top: '40px'},
        {top: '50px'},
        {top: '60px', width: '50px'},
        {top: '70px'},
        {top: '80px'},
        {top: '90px'},
        {top: '100px', width: '100px'}]; //B, pacedB
    var effectsToTest = createEffectsToTest(keyframes, 'paced(width)');
    var expected = computeOffset(keyframes, 'paced', 'width');

    effectsToTest.forEach(function(effect, index) {
        expected.forEach(function(keyframe, i) {
            testKeyframe(effect.getFrames()[i], keyframe, 'Effect ' + index + ', frame ' + i);
        });
    });
}, 'Test that spacing in paced mode algorithm: pacedB == B');

test(function() {
    var keyframes = [{top: '0px', width: '10px'}, // A, pacedA
        {top: '10px'},
        {top: '20px'},
        {top: '30px'},
        {top: '40px'},
        {top: '50px', width: '100px'}]; //B, pacedB
    var effectsToTest = createEffectsToTest(keyframes, 'paced(width)');
    var expected = computeOffset(keyframes, 'paced', 'width');

    effectsToTest.forEach(function(effect, index) {
        expected.forEach(function(keyframe, i) {
            testKeyframe(effect.getFrames()[i], keyframe, 'Effect ' + index + ', frame ' + i);
        });
    });
}, 'Test that spacing in paced mode algorithm: pacedA == A and pacedB == B');


test(function() {
    var keyframes = [{top: '0px'},
        {top: '10px'},
        {top: '20px'},
        {top: '30px'},
        {top: '40px'},
        {top: '100px'}];
    var spacing = 'unsupportedValue';
    var effectsToTest = createEffectsToTest(keyframes, spacing);
    var expected = computeOffset(keyframes, spacing);

    effectsToTest.forEach(function(effect, index) {
        assert_equals(effect.spacing, spacing,
            'Wrong value of KeyframeEffect.spacing attribute');
        expected.forEach(function(keyframe, i) {
            testKeyframe(effect.getFrames()[i], keyframe, 'Effect ' + index + ', frame ' + i);
        });
    });
}, 'Test that all other spacing values are treated as \'distribute\' for ' +
    'the purpose of animation model calculations.');
</script>
</body>
